+++ src/components/app/app.jsx
@@ -1,9 +1,8 @@
 import React from "react";
 import MainScreen from "../main-screen/main-screen.jsx";
 
-const App = () => {
-  return <MainScreen
-  />;
+const App = (props) => {
+  return <MainScreen {props.descriptions}/>;
 };
 
 export default App;
